+++ src/views/index.js
@@ -21,6 +21,7 @@ export { default as listItem } from './list-item.js';
 export { default as lists } from './lists.js';
 export { default as menu } from './menu.js';
 export { default as menuItem } from './menu-item.js';
+export { default as navButton } from './nav-button.js';
 export { default as popup } from './popup.js';
 export { default as section } from './section.js';
 export { default as select } from './select.js';