+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Improved error handling in data prepare handler
 - Changed location path in signature details popup to use `[index]` instead of `pick(index)`
 - Fixed patching for `prismjs@^1.21.0`
+- Reworked navigation panel and introduced `Widget#nav` API
+- Removed `Widget#addBadge()` method (use `Widget#nav` API instead)
 - Changes in views:
     - `button`
         - Added support for `href` and `external` values in `data`, which ignores when button is disabled or `onClick` is specified